refactor(agents): moved server deps from client
