+++ apps/platform/src/components/previews/space-preview.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import type { selectSpaceSchema } from "@curiositi/db/schema";
+import type { selectSpaceSchema } from "@curiositi/share/db-schemas";
 import SpaceDialog from "@platform/components/dialogs/space-dialog";
 import ConfirmDialog from "@platform/components/dialogs/confirm-dialog";
 import {
