+++ packages/db/src/schema.ts
@@ -9,7 +9,12 @@ import {
 	text,
 	timestamp,
 } from "drizzle-orm/pg-core";
-import { z } from "zod";
+import {
+	selectSpaceSchema,
+	createSpaceSchema,
+	selectFileSchema,
+	createFileSchema,
+} from "@curiositi/share/db-schemas";
 
 export const createTable = pgTableCreator((name) => `curiositi_${name}`);
 
@@ -274,27 +279,8 @@ export const spaces = createTable(
 	]
 );
 
-export const selectSpaceSchema = z.object({
-	id: z.string().uuid(),
-	name: z.string(),
-	description: z.string().nullable(),
-	icon: z.string().nullable(),
-	organizationId: z.string(),
-	parentSpaceId: z.string().uuid().nullable(),
-	createdAt: z.date(),
-	updatedAt: z.date().nullable(),
-});
-
-export const createSpaceSchema = z.object({
-	id: z.string().uuid().optional(),
-	name: z.string(),
-	description: z.string().nullable().optional(),
-	icon: z.string().nullable().optional(),
-	organizationId: z.string(),
-	parentSpaceId: z.string().uuid().nullable().optional(),
-	createdAt: z.date().optional(),
-	updatedAt: z.date().nullable().optional(),
-});
+// Re-export schemas from share package for consistency
+export { selectSpaceSchema, createSpaceSchema };
 
 export const fileStatusEnum = pgEnum("file_status", [
 	"pending",
@@ -335,35 +321,8 @@ export const files = createTable(
 	]
 );
 
-export const selectFileSchema = z.object({
-	id: z.uuid(),
-	name: z.string(),
-	path: z.string(),
-	size: z.number().int(),
-	type: z.string(),
-	organizationId: z.string(),
-	uploadedById: z.string(),
-	status: z.enum(["pending", "processing", "completed", "failed"]),
-	tags: z.unknown().nullable(),
-	processedAt: z.date().nullable(),
-	createdAt: z.date(),
-	updatedAt: z.date().nullable(),
-});
-
-export const createFileSchema = z.object({
-	id: z.uuid().optional(),
-	name: z.string(),
-	path: z.string(),
-	size: z.number().int(),
-	type: z.string(),
-	organizationId: z.string(),
-	uploadedById: z.string(),
-	status: z.enum(["pending", "processing", "completed", "failed"]).optional(),
-	tags: z.unknown().nullable().optional(),
-	processedAt: z.date().nullable().optional(),
-	createdAt: z.date().optional(),
-	updatedAt: z.date().nullable().optional(),
-});
+// Re-export schemas from share package for consistency
+export { selectFileSchema, createFileSchema };
 
 export const fileContents = createTable(
 	"file_contents",
